string.[ch]

+++ include/lib/debug.h
@@ -8,7 +8,7 @@
 #ifndef DEBUG_H
 #define DEBUG_H
 
-#include "../core/core.h"
+#include <lib/core.h>
 
 void _debug(char *,char *, uint32);
 void _panic(char *, char*, uint32);
@@ -17,6 +17,6 @@ void trace_stack(uint32);
 
 #define DEBUG(x) _debug(x, __FILE__, __LINE__);
 #define PANIC(x) _panic(x, __FILE__, __LINE__);
-#define ASSERT(x) _assert(x, __FILE__, __LINE__);
+#define ASSERT(x) _assert(((int32)x), __FILE__, __LINE__);
 
 #endif
